Create HyperGalacticWeatherNFT.sol
This smart contract enables the creation of unique NFTs that display abnormal weather conditions on Hyper Galactic planets using the ERC-721 standard. Each NFT is created with a unique token ID and the weather data is stored within the smart contract.

The smart contract is an ERC-721 token named "HyperGalacticWeatherNFT". It uses the "Counters" feature to determine the token IDs for the NFTs. Additionally, a "WeatherData" struct is defined within the contract to store the weather data for each NFT.

The "mintWeatherNFT" function is used to create NFTs. When called by a user, this function creates a new NFT and transfers it to the user's wallet. The token ID for the new NFT is automatically determined using the "Counters" feature. A unique weather data is also generated for the new NFT using random numbers.

The "getWeatherData" function is used to access the weather data for an NFT. It takes a token ID as input and returns the weather data for the NFT with that token ID.

This smart contract enables the creation of unique NFTs that display abnormal weather conditions on Hyper Galactic planets. These NFTs can be listed for sale in an NFT marketplace or used to add to collections. The smart contract provides many features such as automatic transfer of NFTs and uses the blockchain network for their transfer and usage.